+++ lib/doorkeeper/models/access_grant_mixin.rb
@@ -9,7 +9,7 @@ module Doorkeeper
     include Models::Revocable
     include Models::Accessible
     include Models::Orderable
-    include Models::Hashable
+    include Models::SecretStorable
     include Models::Scopes
 
     # never uses pkce, if pkce migrations were not generated
@@ -32,12 +32,6 @@ module Doorkeeper
       #
       def by_token(token)
         find_by_plaintext_token(:token, token)
-      end
-
-      # We want to perform secret hashing whenever the user
-      # enables the configuration option +hash_token_secrets+
-      def perform_secret_hashing?
-        Doorkeeper.configuration.hash_token_secrets?
       end
 
       # Revokes AccessGrant records that have not been revoked and associated
@@ -100,6 +94,20 @@ module Doorkeeper
 
       def pkce_supported?
         new.pkce_supported?
+      end
+
+      ##
+      # Determines the secret storing transformer
+      # Unless configured otherwise, uses the plain secret strategy
+      def secret_strategy
+        ::Doorkeeper.configuration.token_secret_strategy
+      end
+
+      ##
+      # Determine the fallback storing strategy
+      # Unless configured, there will be no fallback
+      def fallback_secret_strategy
+        ::Doorkeeper.configuration.token_secret_fallback_strategy
       end
     end
   end