+++ lib/doorkeeper/orm/active_record/access_grant.rb
@@ -27,16 +27,17 @@ module Doorkeeper
 
     before_validation :generate_token, on: :create
 
-    # Keep a reference to the generated token during generation
-    # of this access grant. The actual token may be mapped by
-    # the configuration hasher and may not be available in plaintext.
+    # We keep a volatile copy of the raw token for initial communication
+    # The stored refresh_token may be mapped and not available in cleartext.
     #
-    # If hash tokens are enabled, this will return nil on fetched tokens
+    # Some strategies allow restoring stored secrets (e.g. symmetric encryption)
+    # while hashing strategies do not, so you cannot rely on this value
+    # returning a present value for persisted tokens.
     def plaintext_token
-      if perform_secret_hashing?
-        @raw_token
+      if secret_strategy.allows_restoring_secrets?
+        secret_strategy.restore_secret(self, :token)
       else
-        token
+        @raw_token
       end
     end
 
@@ -48,7 +49,7 @@ module Doorkeeper
     #
     def generate_token
       @raw_token = UniqueToken.generate
-      self.token = hashed_or_plain_token(@raw_token)
+      secret_strategy.store_secret(self, :token, @raw_token)
     end
   end
 end