+++ NEWS.md
@@ -7,6 +7,12 @@ User-visible changes worth mentioning.
 
 ## master
 
+- [#1208] Unify hashing implementation into secret storing strategies
+
+  **[IMPORTANT]**: If you have been using the master branch of doorkeeper with bcrypt in your Gemfile.lock,
+  your application secrets have been hashed using BCrypt. To restore this behavior, use the initializer option
+  `use_application_hashing using: 'Doorkeeper::SecretStoring::BCrypt`.
+
 - [#1216] Add nil check to `expires_at` method.
 - [#1215] Fix deprecates for Rails 6.
 - [#1214] Scopes field accepts array.